+++ omc3/model/accelerators/lhc.py
@@ -101,6 +101,8 @@ from omc3.model.constants import (
     MACROS_DIR,
     MODIFIER_TAG,
 )
+from omc3.correction.constants import ORBIT_DPP
+
 from omc3.utils import logging_tools
 
 if TYPE_CHECKING:
@@ -398,6 +400,7 @@ class Lhc(Accelerator):
         madx_script += "\n! ---- Call optics and other modifiers ----\n"
 
         if self.modifiers is not None:
+            # if modifier is an absolute path, go there, otherwise use the path refers from model_dir
             madx_script += "".join(
                 f"call, file = '{self.model_dir / modifier}'; {MODIFIER_TAG}\n"
                 for modifier in self.modifiers
@@ -486,12 +489,48 @@ class Lhc(Accelerator):
                 "has to be specified (--lhcmode option missing?)."
             )
 
-    def get_update_correction_script(self, outpath: Path | str, corr_files: Sequence[Path | str]) -> str:
+    def get_update_correction_script(self, outpath: Path | str, corr_files: Sequence[Path | str], update_dpp: bool = False) -> str:
         madx_script = self.get_base_madx_script()
-        for corr_file in corr_files:
+
+        # First set the dpp to the value in the accelerator model
+        madx_script += f"{ORBIT_DPP} = {self.dpp};\n"
+
+        for corr_file in corr_files:  # Load the corrections, can also update ORBIT_DPP
             madx_script += f"call, file = '{str(corr_file)}';\n"
-        madx_script += f"exec, do_twiss_elements(LHCB{self.beam}, '{str(outpath)}', {self.dpp});\n"
+        
+        if update_dpp: # If we are doing orbit correction, we need to ensure that a correct and a match is done (in get_update_deltap_script)
+            madx_script += self.get_update_deltap_script(deltap=ORBIT_DPP)
+
+        madx_script += f'exec, do_twiss_elements(LHCB{self.beam}, "{str(outpath)}", {ORBIT_DPP});\n'
         return madx_script
+    
+    def get_update_deltap_script(self, deltap: float | str) -> str:
+        if not isinstance(deltap, str):
+            deltap = f"{deltap:.15e}"
+
+        madx_script = (
+            f"twiss, deltap={deltap};\n"
+            "correct, mode=svd;\n\n"
+            
+            "! The same as match_tunes, but include deltap in the matching\n"
+            f"exec, find_complete_tunes({self.nat_tunes[0]}, {self.nat_tunes[1]}, {self.beam});\n"
+            f"match, deltap={deltap};\n"
+        ) # Works better when split up
+        madx_script += "\n".join([f"vary, name={knob};" for knob in self.get_tune_knobs()]) + "\n"
+        madx_script += (
+            "constraint, range=#E, mux=total_qx, muy=total_qy;\n"
+            "lmdif, tolerance=1e-10;\n"
+            "endmatch;\n"
+        )
+        return madx_script
+    
+    def get_tune_knobs(self) -> tuple[str, str]:
+        if self._uses_run3_macros():
+            return f"dQx.b{self.beam}_op", f"dQy.b{self.beam}_op"
+        elif self._uses_ats_knobs():
+            return f"dQx.b{self.beam}", f"dQy.b{self.beam}"
+        else:
+            return f"KQTD.B{self.beam}", f"KQTF.B{self.beam}"
 
     # Private Methods ##########################################################